// src/pages/product/type.tsx
export interface Product {
    ROPAN: string;
    ID_ROPAN: string;
    IMAGEN:string;
    LINK: string;
    DETALLEN: string;
    PRECION: number;
    
} 
export interface Productd {
    ROPAD: string;
    ID_ROPAD: string;
    IMAGED:string;
    LINKD: string;
    DETALLED: string;
    PRECIOD: number;
    
} 

export interface Productc {
    ROPAC: string;
    ID_ROPAC: string;
    IMAGEC:string;
    LINKC: string;
    DETALLEC: string;
    PRECIOC: number;
    
} 